# Optimize TestApiStatsNetworkStats and TestApiStatsNetworkStatsVersioning