Before Tests results:

    DockerSuite.TestApiStatsNetworkStatsVersioning                  14.699s
    DockerSuite.TestApiStatsNetworkStats                            13.200s

After Tests results:

    PASS: docker_api_stats_test.go:82: DockerSuite.TestApiStatsNetworkStats	6.894s
    PASS: docker_api_stats_test.go:128: DockerSuite.TestApiStatsNetworkStatsVersioning	9.521s

the results should be accepted.

Related issue #19425 
Signed-off-by: Wen Cheng Ma <wenchma@cn.ibm.com>